requirements.txt
nltk==3.6.1
tensorflow==2.5.0rc0
py_currency_converter==1.2.0
streamlit==0.80.0
joblib==1.0.1
pandas==1.2.3
plotly==4.14.3
matplotlib==3.3.4
yfinance==0.1.55
numpy==1.19.5
keras_tuner==1.0.2
beautifulsoup4==4.9.3
Pillow==8.2.0
scikit_learn==0.24.1
